Start a profiling mod
